Componetized createRoomModal
